requirements.txt
google-generativeai==0.8.3
streamlit==1.38.0
flask==3.0.3
python-dotenv==1.0.1
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
oauth2client==4.1.3
